requirements_Version2.txt
# Airflow
apache-airflow==2.8.0
apache-airflow-providers-apache-spark==4.6.0

# Spark 3.4.1
pyspark==3.4.1

# MLflow
mlflow==2.10.0

# Additional dependencies
pandas==2.0.3
numpy==1.24.3
pyarrow==12.0.1

# Hive dependencies
thrift==0.16.0